# api/openapi.yaml
openapi: 3.1.0
info:
  title: Trace HTTP API
  version: 1.0.0
  summary: Local-first JSON API for Trace (opt-in via `trace serve`)
  description: |
    Versioned HTTP/JSON contract for Trace core. Default bind is loopback
    (`127.0.0.1`). Handlers and the browser GUI are adapters only (Law 19) —
    they call the canonical Go library / `.trace/` SQLite; they must not fork
    business logic or create a second source of truth.

    **Non-goals (explicit):**
    - No `POST /rpc` / MCP tools/call mirror for the browser (MCP stays stdio-local)
    - No default unbounded / full-graph dump (including seed export body over HTTP)
    - No multi-tenant hosting, OAuth, or billing routes in this phase
    - No CORS `Access-Control-Allow-Origin: *` (deny by default; same-origin SPA)

    Wave tags: `x-trace-wave: p0 | p1 | defer`. GUI relevance: `x-trace-gui: true`.
  contact:
    name: Trace
  license:
    name: See repository LICENSE
servers:
  - url: http://127.0.0.1:7432
    description: Local opt-in `trace serve` (placeholder port; CLI may choose another)
...
security:
  - {}
  - bearerAuth: []
paths:
...
  /v1/tasks:
    get:
      tags: [tasks]
      operationId: listTasks
      summary: List tasks (mirrors `trace tasks` / `trace_tasks`)
      x-trace-wave: p0
      x-trace-gui: true
      parameters:
        - name: goal_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Optional goal UUID filter
        - name: work_state
          in: query
          required: false
          schema:
            type: string
          description: Optional work_state filter
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque pagination cursor (optional; S03 may start unpaginated)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/tasks/{task_id}:
    get:
      tags: [tasks]
      operationId: getTask
      summary: Get one task by id
      x-trace-wave: p0
      x-trace-gui: true
      parameters:
        - $ref: "#/components/parameters/TaskIdPath"
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskRow"
        "404":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
  /v1/reviews:
    get:
      tags: [reviews]
      operationId: listReviews
      summary: List reviews
      x-trace-wave: p1
      x-trace-gui: true
      parameters:
        - name: task_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Review list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReviewSummary"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [reviews]
      operationId: createReview
      summary: Create review (`trace review` create)
      x-trace-wave: p1
      x-trace-gui: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                body:
                  type: string
                task_id:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewSummary"
        default:
          $ref: "#/components/responses/Error"
  /v1/reviews/{review_id}:
    get:
      tags: [reviews]
      operationId: getReview
      summary: Get review
      x-trace-wave: p1
      x-trace-gui: true
      parameters:
        - name: review_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewSummary"
        default:
          $ref: "#/components/responses/Error"
...
  /v1/changes:
    get:
      tags: [changes]
      operationId: listChanges
      summary: Change listing
      x-trace-wave: p1
      x-trace-gui: true
      responses:
        "200":
          description: Changes
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: "#/components/responses/Error"
  /v1/regressions:
    get:
      tags: [regressions]
      operationId: listRegressions
      summary: Regression listing
      x-trace-wave: p1
      x-trace-gui: true
      responses:
        "200":
          description: Regressions
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Required when `trace serve` binds non-loopback (`--allow-remote`).
        On loopback, authentication may be omitted (loopback-trust). Optional
        loopback token is not required for MVP.
  parameters:
    TaskIdPath:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable code (e.g. VALIDATION_ERROR)
            message:
              type: string
            details:
              description: Optional structured context
              nullable: true
...
    TaskRow:
      type: object
      required: [id, title, work_state]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        work_state:
          type: string
        goal_id:
          type: string
          format: uuid
          nullable: true
        body:
          type: string
          nullable: true
    TaskListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TaskRow"
        next_cursor:
          type: string
          nullable: true
...
    ReviewSummary:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
          nullable: true
        result:
          type: string
          enum: [PASS, FAIL, UNCERTAIN]
          nullable: true
        task_id:
          type: string
          format: uuid
          nullable: true